moved an id to the right shape query selectors
one problem that we have now is that is we clone a shape and we apply a
filter to its parent then
the clone gets the filter applied to. one way around it is to create a
<def> of the shape to be cloned
and then make two copies and apply the filter to each separately.
however there is no way to target sub elements
withing a clone so the filter gets added to all subshapes. a way around
it is to create two shapes per letter
one the red and one the tube and position them layered. then we have
autonomy when applying filters and colours